requirements.txt
fastapi
pillow
deepface
tf-keras
cv2
numpy